Create delete_duplicates.sql
Pretty straight forward, sorts by "select distinct" which is the columns you want to check matching on.  So this query will look at every row that has the same data in BOTH those rows to deem a duplicate, and then "pk_id" is the PRIMARY KEY that it'll decide to get rid of in the dupe.